% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fullEM}
\alias{fullEM}
\title{Doing the full EM Algorithm}
\usage{
fullEM(x, y, weights, sd, tolerance)
}
\arguments{
\item{x}{The Numeric Matrix of X values for the models}

\item{y}{A vector of values for each observation}

\item{weights}{The vector of weights for each model}

\item{sd}{Standard deviation}

\item{tolerance}{The level of chnage between iterations where the model will stop}
}
\description{
Doing the full EM Algorithm
}
