% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{w_hat}
\alias{w_hat}
\title{Solving Equation 3 of the EM algorithm by computing our weights}
\usage{
w_hat(z)
}
\arguments{
\item{z}{The matrix of prodicted z values}
}
\description{
Solving Equation 3 of the EM algorithm by computing our weights
}
