% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{z_hat}
\alias{z_hat}
\title{Solving Equation 2 of the EM Algorithm for Z_Hat}
\usage{
z_hat(x, y, weights, sd)
}
\arguments{
\item{x}{The Numeric Matrix of X values for the models}

\item{y}{A vector of values for each observation}

\item{weights}{The vector of weights for each model}

\item{sd}{Standard deviation}
}
\description{
Solving Equation 2 of the EM Algorithm for Z_Hat
}
